Fixed a binding declaration and updated Business delete person to handle null or new record correctly.

+++ Model/Business.cs
@@ -23,6 +23,8 @@ namespace MVVMApplication.Model
 
         internal void Delete(Person personToDelete)
         {
+            if(personToDelete == null || personToDelete.Id == 0) return;
+
             Person foundPerson = _dbContext.Person.Find(new object[] { personToDelete.Id });
             _dbContext.Person.Remove(foundPerson);
             _dbContext.SaveChanges();
